pyproject.toml
[project]
name = "estatya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "aiosqlite>=0.22.1",
    "python-jose[cryptography]>=3.5.0",
    "argon2-cffi>=25.1.0",
    "httpx>=0.28.1",
]
[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["src/tests"]